+++ crates/papyrus_gateway/src/state.rs
@@ -1,11 +1,12 @@
 use std::collections::HashMap;
 
-use papyrus_storage::state::data::ThinStateDiff as papyrus_storage_ThinStateDiff;
 use serde::{Deserialize, Serialize};
 use starknet_api::block::BlockHash;
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, GlobalRoot, Nonce};
 use starknet_api::hash::StarkFelt;
-use starknet_api::state::{EntryPoint, EntryPointType, StorageKey};
+use starknet_api::state::{
+    EntryPoint, EntryPointType, StorageKey, ThinStateDiff as starknet_api_ThinStateDiff,
+};
 
 const CONTRACT_CLASS_VERSION: &str = "0.1.0";
 
@@ -27,8 +28,8 @@ pub struct ThinStateDiff {
     pub replaced_classes: Vec<ReplacedClasses>,
 }
 
-impl From<papyrus_storage_ThinStateDiff> for ThinStateDiff {
-    fn from(diff: papyrus_storage_ThinStateDiff) -> Self {
+impl From<starknet_api_ThinStateDiff> for ThinStateDiff {
+    fn from(diff: starknet_api_ThinStateDiff) -> Self {
         Self {
             deployed_contracts: Vec::from_iter(
                 diff.deployed_contracts
